Updated javadoc, code cleanup, auth change based on parameter.
