+++ marklogic-rdf4j/src/main/java/com/marklogic/semantics/rdf4j/client/TripleWriteCache.java
@@ -32,7 +32,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 /**
- * Created by jfuller on 11/2/16.
+ * Write cache to optimize performance by batching the requests.
  */
 public class TripleWriteCache extends TripleCache {
 
@@ -46,7 +46,7 @@ public class TripleWriteCache extends TripleCache {
         super(client, cacheSize);
     }
     /**
-     * flushes the cache, writing triples as graph
+     * Flushes the cache to the server, writing triples as graph.
      *
      * @throws MarkLogicRdf4jException
      */
