+++ marklogic-rdf4j/src/main/java/com/marklogic/semantics/rdf4j/utils/Util.java
@@ -1,13 +1,51 @@
 package com.marklogic.semantics.rdf4j.utils;
 
+import com.marklogic.client.DatabaseClient;
+import com.marklogic.client.DatabaseClientFactory;
 import org.eclipse.rdf4j.model.Value;
 import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
 import org.eclipse.rdf4j.rio.RDFFormat;
+
+import java.io.IOException;
+import java.security.KeyManagementException;
+import java.security.UnrecoverableKeyException;
+import java.security.cert.CertificateException;
 
 public class Util {
     private static Util util = null;
     private Util(){
 
+    }
+
+    /**
+     * Authentication enumerates the methods for verifying a user and
+     * password with the database.
+     */
+    public enum Authentication {
+        /**
+         * Minimal security unless used with SSL.
+         */
+        BASIC,
+        /**
+         * Moderate security without SSL.
+         */
+        DIGEST,
+        /**
+         * Authentication using Kerberos.
+         */
+        KERBEROS,
+        /**
+         * Authentication using Certificates;
+         */
+        CERTIFICATE;
+        /**
+         * Returns the enumerated value for the case-insensitive name.
+         * @param name	the name of the enumerated value
+         * @return	the enumerated value
+         */
+        static public Authentication valueOfUncased(String name) {
+            return Authentication.valueOf(name.toUpperCase());
+        }
     }
 
     /**
@@ -34,7 +72,7 @@ public class Util {
     /**
      * Public utility method that skolemizes blank nodes (BNode).
      * @param s the blank node to be skolemized.
-     * @return for BNode return skolemized BNode or else the node itself.
+     * @return for BNode returns skolemized BNode or else the node itself.
      */
     public Value skolemize(Value s) {
         if (s instanceof org.eclipse.rdf4j.model.BNode) {
@@ -53,5 +91,42 @@ public class Util {
         return dataFormat.equals(RDFFormat.TURTLE) || dataFormat.equals(RDFFormat.RDFXML) || dataFormat.equals(RDFFormat.TRIG)
             || dataFormat.equals(RDFFormat.NQUADS) || dataFormat.equals(RDFFormat.NTRIPLES) || dataFormat.equals(RDFFormat.RDFJSON)
                 || dataFormat.equals(RDFFormat.N3);
+    }
+
+    /**
+     * Public utility that returns DatabaseClient based on auth
+     * @return DatabaseClient
+     */
+    public DatabaseClient getClientBasedOnAuth(String host, int port, String user, String password, String auth, String... cert) throws UnrecoverableKeyException, CertificateException, KeyManagementException, IOException {
+        Authentication type;
+        String certFile;
+        String certPassword;
+
+        if(auth != null)
+        {
+            type = Authentication.valueOfUncased(auth);
+            certFile = cert.length > 0 ? cert[0] : "";
+            certPassword = cert.length > 1 ? cert[1] : "";
+
+            if(type == Authentication.BASIC)
+            {
+                return DatabaseClientFactory.newClient(host, port, new DatabaseClientFactory.BasicAuthContext(user, password));
+            }
+            else if(type == Authentication.DIGEST)
+            {
+                return DatabaseClientFactory.newClient(host, port, new DatabaseClientFactory.DigestAuthContext(user, password));
+            }
+            else if(type == Authentication.KERBEROS)
+            {
+                return DatabaseClientFactory.newClient(host, port, new DatabaseClientFactory.KerberosAuthContext());
+            }
+            else if(type == Authentication.CERTIFICATE)
+            {
+                //TODO: change parameters
+                return DatabaseClientFactory.newClient(host, port, new DatabaseClientFactory.CertificateAuthContext(certFile, certPassword));
+            }
+        }
+
+        return null;
     }
 }
